common/requirements.txt
# Requisitos generales para FastAPI
fastapi==0.110.0

# Seguridad y autenticación (opcional, si usas JWT por ejemplo)
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pyjwt
